requirements.txt
fastapi
uvicorn
google-cloud-vision
pillow
python-multipart